backend/requirements.txt
fastapi
uvicorn[standard]
python-pptx
python-docx
pdf2docx
deep-translator
